+++ NEWS.md
@@ -1,7 +1,10 @@
 # r2dii.analysis (development version)
 
+* New internal data `loanbook_stable` and `region_isos_stable` make regression
+  tests more stable (#227).
+
 * Hyperlinks on the "Get Started" tab of the website now points to correct links 
-  (#222 @apmanning)
+  (#222 @apmanning).
 
 * Depend on dplyr >= 0.8.5, explicitly. We commit to this version because the
   newer dplyr 1 is still relatively new, and represents a major change which